requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt setup.py
#
attrs==18.2.0
python-dateutil==2.8.0
six==1.12.0               # via python-dateutil
